feat(208-01): propagate type narrowing through AdvanceModule, GameOverModule, DecimatorModule
- uint48→uint32 for all day-index vars, params, locals, events, constants
- Direct lootboxRng/gameOver/presale variable access → packed slot helpers (_lrRead/_lrWrite, _goRead/_goWrite, _psRead/_psWrite)
- claimablePool arithmetic cast to uint128 with safety comments
- All remaining uint48 are timestamps (correct)

+++ contracts/modules/DegenerusGameDecimatorModule.sol
@@ -363,7 +363,7 @@ contract DegenerusGameDecimatorModule is DegenerusGamePayoutUtils {
         _creditClaimable(account, ethPortion);
 
         // Lootbox portion is no longer claimable ETH; remove from reserved pool.
-        claimablePool -= lootboxPortion;
+        claimablePool -= uint128(lootboxPortion); // Safe: lootboxPortion is a fraction of claimablePool, fits uint128
         _awardDecimatorLootbox(account, lootboxPortion, rngWord);
     }
 
@@ -613,8 +613,8 @@ contract DegenerusGameDecimatorModule is DegenerusGamePayoutUtils {
     // Terminal Decimator Constants
     // -------------------------------------------------------------------------
 
-    uint48 private constant TERMINAL_DEC_IDLE_TIMEOUT_DAYS = 365;
-    uint48 private constant TERMINAL_DEC_DEATH_CLOCK_DAYS = 120;
+    uint32 private constant TERMINAL_DEC_IDLE_TIMEOUT_DAYS = 365;
+    uint32 private constant TERMINAL_DEC_DEATH_CLOCK_DAYS = 120;
 
     // -------------------------------------------------------------------------
     // Terminal Decimator Burn Tracking
@@ -875,8 +875,8 @@ contract DegenerusGameDecimatorModule is DegenerusGamePayoutUtils {
 
     /// @dev Calculate days remaining on death clock using day-index arithmetic. Returns 0 if expired.
     function _terminalDecDaysRemaining() private view returns (uint256) {
-        uint48 currentDay = _simulatedDayIndex();
-        uint48 psd = purchaseStartDay;
+        uint32 currentDay = _simulatedDayIndex();
+        uint32 psd = purchaseStartDay;
         uint256 deadlineDay = uint256(psd) +
             (
                 level == 0
